Fixes to some HA scenarios.
